Changed wrangle_plot_data()
Rewrote a large part of this function around the single and pipeline mode. The flow is now more consistent with the other pipeline functions.

Modes are set by whether a project name or a file path is supplied; the former should imply plot_fluxxer() called it in pipeline mode, and it will scan the folder structure (whether or not there's >1 project is not important). In single file mode, no looping happens and the file is simply read in directly with no checks.

Refactoring is now handled by refactor_reps(), and it's agnostic about whether levels are actually a sensible approach, because even if there's only one data point, goodOrder will work fine.

Also implemented a check that the highest CI upper and the lowest CI lower are a difference of more than 1 power of 10. This is later passed as a flag to plot_mutrates().