Changed plot_fluxxer()
Rewrote some of the function to fit the mode system. The check for folder structure now only happens in pipeline mode.

plot_mutrates() is now called with a filename or project prefix depending on single or standard pipeline mode respectively. An additional argument (the "log" flag) is now part of the call.

export_mut_plot() is now in place, but not tested. Fairly sure nothing's changed since it was committed, but good luck finding that commit.