Implemented overwriting option
Functions now allow selecting overwrite = TRUE. If so, the outputs will be dumped into the appropriate folder regardless of its contents. For plots, this is implemented as a new mode in prep_export(mode = "plots"), which will check if the output folder has plots in order to warn in the usual way.