Created test_logticks()
Made the annotation_logticks() tester a function so it can be called by both modes in wrangle_plot_data(). test_logticks() simply take the data and runs the test, returning logMode().

This is actually fine for standard mode too, because it is passed combinedData, and will only trigger the logMode change if ALL the data meets the error criteria.