Updated plot_fluxxer()
Changed project name grepping code to use inputPath, so it only has to be defined once. wrangle_plot_data is called using inputPath as well, which it uses to parse the filenames using projectName.

Changed inputPath to be /analyzed/ since this is the actual source of plotting data, so it's more appropriate as a check for the previous pipeline step.

Fixed prefix extraction in single mode to use basename(file) instead of file directly.