Created extract_mutrates()
A simple function to extract pooled mutation rates from multiple project files. This used to be in get_mutation_rates(), built for use with pkg tapedeck; the intention was to make mutation rate versions easier to track, instead of e.g. copy pasting values from the CSVs.

The relevant mode has been added to prep_export(), this time checking for files starting with "mutrate_", and exporting to /analyzed/.

This doesn't have a single-file mode, because it should only be used on pooled data and I don't want to implement some kind of internal check for this on the supplied file.